refactor: detail page 분리
